non working jacobian calculation and test bed. has issues

+++ kinematics-fixed/kinematics-fixed/utils.cpp
@@ -9,11 +9,11 @@ void print_mat4_32b(mat4_32b_t m)
 
 		for (int c = 0; c < 4; c++)
 		{
-			float v;
+			double v;
 			if (c == 3 && r < 3)
-				v = ((float)m.m[r][c]) / ((float)(1 << 16));
+				v = ((double)m.m[r][c]) / ((double)(1 << 16));
 			else
-				v = ((float)m.m[r][c]) / ((float)(1 << 21));
+				v = ((double)m.m[r][c]) / ((double)(1 << 21));
 			printf("%f ", v);
 		}
 		printf("\n");
